requirements.txt
line-bot-sdk
flask
bs4
selenium
requests
lxml
psycopg2
